update toast.js, selectize.jsをdefault.jsに統合。ファイル削除済み
複数回jsファイルを読み込むことを避けるため、default.jsを作成してapplication.jsで読み込むように修正。
