+++ app/frontend/entry/application.js
@@ -19,6 +19,7 @@ const images = require.context("../images", true);
 // const imagePath = (name) => images(name, true)
 
 // import "../fonts/memoir-round.otf"
-import "bootstrap";
+// import "bootstrap";
 import "@fortawesome/fontawesome-free/js/all";
+import "../js/default";
 import "../scss/application";
