--- app/frontend/entry/default.js
@@ -1,12 +1,0 @@
-// import HeaderDropdownMenu from "../js/components/header_dropdown_menu";
-
-// document.addEventListener('DOMContentLoaded', function (event) {
-//   if (document.getElementById("header")) {
-//     new vue({
-//       el: "#header",
-//       components: {
-//         "header-dropdown-menu": HeaderDropdownMenu
-//       }
-//     })
-//   }
-// })
