--- app/frontend/entry/toast.js
@@ -1,12 +1,0 @@
-window.addEventListener('load', function () {
-  const type = document.getElementById("toast_type").value;
-  const message = document.getElementById("toast_message").value;
-  Swal.fire({
-    type: type,
-    title: message,
-    toast: true,
-    position: 'bottom-start',
-    showConfirmButton: false,
-    showCloseButton: true
-  });
-});
