+++ app/frontend/scss/application.scss
@@ -13,6 +13,7 @@ $fa-font-path: "~@fortawesome/fontawesome-free/webfonts";
 @import "~@fortawesome/fontawesome-free/scss/brands";
 @import "trix/dist/trix";
 @import "selectize/dist/css/selectize";
+@import "tippy.js/dist/tippy";
 
 /*===============================
 foundation
