+++ app/views/layouts/article_post.html.slim
@@ -8,4 +8,4 @@ html lang = 'ja'
     - if Rails.env.development?
       = debug(params)
     = render partial: 'shared/flash'
-    = javascript_pack_tag 'selectize'
+    / = javascript_pack_tag 'selectize'
