+++ package.json
@@ -11,7 +11,6 @@
     "axios": "^0.19.0",
     "bootstrap": "^4.3.1",
     "jquery": "^3.4.1",
-    "popper.js": "^1.15.0",
     "ress": "^2.0.1",
     "selectize": "^0.12.6",
     "sweetalert2": "^8.18.3",
